% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_var}
\alias{check_var}
\title{Check Variables}
\usage{
check_var(environmental_variable)
}
\arguments{
\item{environmental_variable}{An environmental variable to check for.}
}
\description{
Check Variables
}
\examples{
\dontrun{
check_var("GOOGLE_FIT_KEY"),
}

}
