% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_api.R
\name{fitness_api}
\alias{fitness_api}
\title{Fitness API}
\usage{
fitness_api(path, token)
}
\arguments{
\item{path}{An endpoint to specifically query. See \url{https://developers.google.com/fit/rest}}

\item{token}{A token from \code{\link{fitness_store_auth}}}
}
\value{
A response object
}
\description{
Fitness API
}
\examples{
\dontrun{
res <- fitness_api("fitness/v1/users/me/sessions", google_token)
}

}
