% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_store_auth.R
\name{fitness_store_auth}
\alias{fitness_store_auth}
\title{get_google_oauth2_token}
\usage{
fitness_store_auth(cache = TRUE)
}
\arguments{
\item{cache}{Should token be stored locally? Defaults to TRUE.}
}
\description{
Authorize access to Google Fitness Store. Access token will be cached
by default.
}
\examples{
\dontrun{
#First set up client id and secret in
# google cloud console
#https://developers.google.com/fit/rest/v1/get-started

fitness_store_auth()

#add some subsequent steps
}

}
