% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sessions.R
\name{get_sessions}
\alias{get_sessions}
\title{Get Sessions}
\usage{
get_sessions()
}
\description{
Get the time interval during which users perform a fitness activity.
More info at \url{https://developers.google.com/fit/rest/v1/using-sessions}
}
\examples{
\dontrun{
get_sessions
}

}
