% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFitDataset.R
\name{GetFitDataset}
\alias{GetFitDataset}
\title{GetFitDataset}
\usage{
GetFitDataset(token, datasource, startTime, endTime)
}
\arguments{
\item{token}{- OAuth 2.0 access token}

\item{datasource}{- Data Stream ID}

\item{startTime}{- Start time for dataset. \code{Date} datatype is required}

\item{endTime}{- End time for dataset. \code{Date} datatype is required}
}
\description{
Retrieves a dataset for a given \code{datasource} and time range defined by
\code{startTime} and \code{endTime}.
Refer to \url{https://developers.google.com/fit/rest/v1/datasets} for full documentation.
}
