% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFitDatasource.R
\name{GetFitDatasource}
\alias{GetFitDatasource}
\title{GetFitDatasource}
\usage{
GetFitDatasource(token, datasource)
}
\arguments{
\item{token}{- OAuth 2.0 access token}

\item{datasource}{- Data Stream ID}
}
\description{
Returns a list with datasource attributes.
Refer to \url{https://developers.google.com/fit/rest/v1/data-types} for full documentation.
}
