% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFitDatasources.R
\name{GetFitDatasources}
\alias{GetFitDatasources}
\title{GetFitDatasources}
\usage{
GetFitDatasources(token)
}
\arguments{
\item{token}{- OAuth 2.0 access token}
}
\description{
\code{GetFitDatasources} returns a dataframe with user's datasources.
}
