% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFitOAuth2Token.R
\name{GetFitOauth2Token}
\alias{GetFitOauth2Token}
\title{GetFitOauth2Token}
\usage{
GetFitOauth2Token(client_id = NULL, client_secret = NULL,
  scopes = NULL, use_oob = F, cache = T)
}
\arguments{
\item{client_id}{Client ID from console.developers.google.com}

\item{client_secret}{Client Secret from console.developers.google.com}

\item{scopes}{Scopes required for requests. List of scopes available on the [Google Fitness REST API Documentation](https://developers.google.com/fit/rest/v1/authorization)}

\item{use_oob}{if FALSE (the default), use a local webserver (http://localhost/) for the OAuth dance. Otherwise, provide a URL to the user and prompt for a validation code}

\item{cache}{A logical value or a string. TRUE (the default) means to cache using the default cache file .httr-oauth, FALSE means don't cache, and NA means to guess using some sensible heuristics. A string means use the specified path as the cache file.}
}
\description{
Retrieves or refreshes an OAuth2 token. Two options must be set or provided as arguments:
\itemize{
\item \code{RGoogleFit.client_id}
\item \code{RGoogleFit.client_secret}
}
Requires setup of API Credentials. A good tutorial for setting up Google REST API credentials can be found here: [How to authenticate using OAuth2 through R](https://biolitika.si/how-to-authenticate-using-oauth2-through-r.html). Simply do these steps for the Google Fitness API.
}
