% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{milliseconds_to_datetime}
\alias{milliseconds_to_datetime}
\title{Milliseconds to Datetime}
\usage{
milliseconds_to_datetime(df)
}
\arguments{
\item{df}{Provide a dataframe from \code{\link{fitness_api}}}
}
\description{
Milliseconds to Datetime
}
