% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{NanosToPOSIXct}
\alias{NanosToPOSIXct}
\title{NanosToPOSIXct}
\usage{
NanosToPOSIXct(nanos)
}
\arguments{
\item{nanos}{- Nanoseconds from epoch}
}
\description{
Converts nanoseconds from epoch (as provided by Google Fit) to POSIXct
}
\examples{
NanosToPOSIXct(1388534400000000000)
}
