% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGoogleFit-package.R
\docType{package}
\name{RGoogleFit-package}
\alias{RGoogleFit-package}
\title{R Interface To Google Fit API}
\description{
Provides interface to Google Fit REST API v1

Sources available at https://github.com/ms32035/RGoogleFit
}

