% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFitOAuth2Token.R
\name{GetFitOauth2Token}
\alias{GetFitOauth2Token}
\title{GetFitOauth2Token}
\usage{
GetFitOauth2Token()
}
\description{
Retrieves or refreshes an OAuth2 token. Two options must be set:
\itemize{
\item \code{RGoogleFit.client_id}
\item \code{RGoogleFit.client_secret}
}
}
